third_party/blink/web_tests/external/wpt/css/selectors/invalidation/option-disabled-when-ancestor-changes-expected.txt

This is a testharness.js-based test.
[FAIL] optgroup :disabled and :has(optgroup:disabled) invalidate when ancestor select disabled toggles
  assert_equals: optgroup color after disabling select expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
[FAIL] optgroup :disabled and :has(optgroup:disabled) invalidate when ancestor fieldset disabled toggles
  assert_equals: optgroup color after disabling fieldset expected "rgb(255, 0, 0)" but got "rgb(0, 0, 0)"
Harness: the test ran to completion.
